feat: add jwt for authorization

+++ main.py
@@ -1,13 +1,14 @@
 from functools import lru_cache
 from typing import Annotated
 
-from fastapi import FastAPI, status, Depends
+from fastapi import FastAPI, status, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 import dependencies
 import models
 import schemas
 import service
+import jwt
 from config import Settings
 from database import get_db
 
@@ -20,8 +21,10 @@ def get_settings():
 
 
 @app.get("/users/{id}", response_model=schemas.User)
-def get_user_by_id(id: int, db: Session = Depends(get_db)):
+def get_user_by_id(id: int, db: Session = Depends(get_db), current_user: dict = Depends(jwt.get_current_user)):
     user = service.get_user_by_id(db, id)
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='User not found')
     return user
 
 
@@ -39,15 +42,23 @@ def create_user(dependency=Depends(dependencies.check_new_user)):
 @app.put("/users/{id}", response_model=schemas.User)
 def update_user(id: int, update_data: schemas.UserUpdateInput, db: Session = Depends(get_db)):
     user = service.get_user_by_id(db, id)
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='User not found')
+
     return service.update_user(db, update_data, user)
 
 
 @app.delete("/users/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_user(id: int, db: Session = Depends(get_db)):
     user = service.get_user_by_id(db, id)
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='User not found')
+
     service.delete_user(db, user)
 
 
-@app.post("/login", response_model=schemas.User)
+@app.post("/login", response_model=schemas.LoginReturn)
 def login(user: models.User = Depends(dependencies.authenticate_user)):
+    access_token = jwt.create_access_token(data={"sub": user.name, "id": user.id})
+    user.token = access_token
     return user
